roles/prepare-ansible-target/tasks/main.yml: Always update the SSH login key.
If an ssh login key has already been set up for the ansibleadmin account but the user under which ansible is run changes their public key (e.g. using ssh-keygen) then running this ansible role will not update the key. There seems to be no way to say "update the key if needed" so for now the workaround is to remove the key and re-add the (currently) correct key.