Finalise the mysql role and add molecule tests.
Note that there is currently no worthwhile infrastructure test.

In the end it proved too much effort to make this role work on Ubuntu and CentOS.
The relevant work has been left in place in case it proves helpful in the future.

The role now installs mariadb, adds an admin user and performs some of the work of securing the installation.
root access has not been altered because it doesn't seem to be possible to login as root with any password.
(With sudo, it is possible to login as root, regardless of password).

tasks/main.yml: Modified the tasks to install mariadb and associated software and to secure the installation.

tasks/secure-mysql.yml: Adds an admin user and removes a few things that mysql_secure_installation would remove.

tasks/warn_password.yml: If appropriate, reminds the invoker that a password should be set.

defaults/main.yml: password and admin user.

vars/CentOS.yml:
vars/Debian.yml:
vars/Ubuntu-16.yml
vars/Ubuntu.yml Various distro-specific variables.

meta/main.yml: galaxy metadata

.yamllint:
molecule/: Molecule infrastructure.