Fix some issues found by ansible-lint in the virtualbox role.
roles/virtualbox/tasks/main.yml: Lint complains that some lines are too long, which was fixed by introducing a fact (variable).
Lint also complained that 'set -o pipefail' needs to be present if pipes are used. The internet suggests that the safe way to do this on Ubuntu is to specify 'executable: /bin/bash'.
The quoting also needs to be tweaked otherwise the revised command line is not correctly understood by bash.

roles/virtualbox/tasks/Debian-Install.yml:
roles/virtualbox/tasks/RedHat-Install.yml: The 'version' variable should be quoeted as "{{ version }}", that is, with a space before and after the variable name.