feat: Update the anki role to install 2.1.20
Since this role was last tested the location of the anki downloads has moved. 2.1.19 no longer seems to be around so switch to 2.1.20 for now.
Eventually this should be updated to a later version.

README.md: Describe the role more fully.

molecule/default/molecule.yml: Switch to using docker for molecule testing: the overhead of vagrant is not necessary for a simple software install. There were issues with some of the debain and centos runs, so knock them out for now. Add Linux Mint 20.1.

tasks/main.yml: Force an "apt-get update" otherwise the ubuntu docker containers will not find build-essentials.

vars/Ubuntu.yml: Fix the name (s/Debian/Ubuntu) and use build-essentials instead of make

vars/main.yml: Change the version number of anki that should be installed