feat: Fix up the image_manipulation role.
ansible requires role names to be restricted to alphanumerics and the underscore: dash is not permitted (although it works). Rename the role.
Getting mupdf-tools to install on CentOS is very, very challenging (unless you trust random repos on the net) so drop CentOS.
pdftk is getting harder and harder to install withoutn using snapd, so drop that.