feat: Install MS Teams as part of the collaboration tools
Added MS Teams.
Made the code more compact.
Added support for CentOS 7/8.

Note that the current release of Skype is broken and fails to install properly (it claims to need libatomic1 - a SUSE library - rather than libatomic.
This will seemingly be fixed in the next release. Everything else work son CentOS so Skype is left in but commented out and the testinfra failure is to be expected on CentOS.